requirements.txt
amqp==5.2.0
ansicon==1.89.0
anyio==4.4.0
arrow==1.3.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
black==24.10.0
blessed==1.20.0
celery==5.4.0
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
coverage==7.6.3
cryptography==43.0.1
daphne==4.1.2
Django==5.1
django-picklefield==3.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
exceptiongroup==1.2.2
Faker==28.0.0
flake8==7.1.1
h11==0.14.0
httptools==0.6.1
hyperlink==21.0.0
idna==3.8
incremental==24.7.2
inflection==0.5.1
iniconfig==2.0.0
jinxed==1.3.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.4.2
mccabe==0.7.0
msgpack==1.0.8
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.48
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.9.0
pyOpenSSL==24.2.1
pytest==8.3.3
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
tomli==2.0.1
Twisted==24.7.0
txaio==23.1.1
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
uvicorn==0.30.6
vine==5.1.0
watchfiles==0.23.0
wcwidth==0.2.13
websocat==1.13.0
websockets==13.0
zope.interface==7.0.3
